Migrate EFS id validator
* Added availability zone attribute to head node networking
* There were no existing tests, still TODO
